Show map also in <Bird> component. Add back location support.
