+++ src/Bird/FindingModal.js
@@ -8,7 +8,7 @@ import {
   PrimaryButton,
   AlertButton
 } from '../common/Button/Button'
-import Map from '../Map/Map2'
+import Map from '../Map/Map'
 import ButtonGroup from '../ButtonGroup/ButtonGroup'
 
 class Modal extends PureComponent {
@@ -105,7 +105,6 @@ export default class FindingModal extends PureComponent {
               onCoordinatesSelected={this.onCoordinatesSelected}
               markerCoordinates={this.state.coordinates}
               bird={this.props.bird}
-              finding={this.props.finding}
             />
           )}
           <SecondaryButton onClick={this.props.onClose}>
