+++ src/Map/Map.module.css
@@ -1,17 +1,18 @@
-.marker {
-  background: red;
-  border: 5px solid white;
-  height: 30px;
-  width: 30px;
-  border-radius: 30px;
+.map {
+  height: 400px;
+  width: 100%;
+  margin-bottom: 20px;
   position: relative;
-  left: -15px;
-  top: -15px;
+}
+
+.roundIcon {
+  border-radius: 100px;
+  border: 5px solid var(--color-white);
 }
 
 .locateButton {
   position: absolute;
-  bottom: 10px;
+  top: 90px;
   left: 10px;
   width: 40px;
   height: 40px;
@@ -23,14 +24,9 @@
   align-items: center;
   color: rgb(90, 90, 90);
   border-radius: 2px;
+  z-index: 1000;
 }
 
 .locateButton:hover {
   color: black;
-}
-
-.map {
-  height: 300px;
-  width: 100%;
-  position: relative;
 }
