--- src/Map/Map2.module.css
@@ -1,9 +1,0 @@
-.map {
-  height: 300px;
-  width: 100%;
-}
-
-.roundIcon {
-  border-radius: 100px;
-  border: 5px solid var(--color-white);
-}
